First working prototype.
First performs a series of preprocessing stages, including removing common, non-location words from tweets.
After these are done, there is a function which takes a chunk of locations (~10000) and checks whether they're inthe tweets. This function is multiprocessed with 16 processes.
The results are then flattened into a final list with the tweet, the location and the cost. These results are dumped to a pickle file for later analysis.

TODO: Need to keep the original tweet alongside the preprocessed tweet for analysis purposes.